chore: move persistence types into a dedicated lib
