+++ Cargo.toml
@@ -15,6 +15,7 @@ license = "Apache-2.0"
 
 [workspace.dependencies]
 amplify = "4.7.0"
+nonasync = "0.1.0"
 bp-std = "0.11.0-beta.7"
 psbt = "0.11.0-beta.7"
 descriptors = "0.11.0-beta.7"
@@ -54,6 +55,7 @@ name = "bpwallet"
 
 [dependencies]
 amplify = { workspace = true }
+nonasync = { workspace = true }
 strict_encoding = "2.7.0-beta.4"
 bp-std = { workspace = true }
 bp-esplora = { workspace = true, optional = true }
@@ -91,3 +93,6 @@ esplora = ["bp-esplora"]
 mempool = ["esplora"]
 fs = ["serde"]
 serde = ["serde_crate", "serde_yaml", "toml", "bp-std/serde"]
+
+[patch.crates-io]
+nonasync = { git = "https://github.com/rust-amplify/amplify-nonasync" }
